+++ examples/post_to_influxdb.py
@@ -18,13 +18,13 @@ $ docker start docker-statsd-influxdb-grafana
 
 InfluxDB: http://localhost:3004/
 
-First create new table: CREATE TABLE tag_data
-Query to return all tag data: SELECT * FROM ruuvitag
-Remove all data: DROP SERIES FROM ruuvitag
+First create new database: CREATE DATABASE ruuvi
+Query to return all tag data: SELECT * FROM ruuvi_measurements
+Remove all data: DROP SERIES FROM ruuvi_measurements
 
 Grafana: http://localhost:3003/ (root/root)
 
-Add datasource (type: InfluxDB url: http://localhost:8086, database: tag_data)
+Add datasource (type: InfluxDB url: http://localhost:8086, database: ruuvi)
 Add new graph to dashboard
 '''
 
@@ -33,20 +33,37 @@ from ruuvitag_sensor.ruuvi import RuuviTagSensor
 
 
 def convert_to_influx(mac, payload):
+    '''
+    Convert data into RuuviCollector naming schme and scale
+
+    returns:
+        Object to be written to InfluxDB
+    '''
+    dataFormat = payload["data_format"] if ('data_format' in payload) else None
+    fields = {}
+    fields["temperature"]               = payload["temperature"] if ('temperature' in payload) else None
+    fields["humidity"]                  = payload["humidity"] if ('humidity' in payload) else None
+    fields["pressure"]                  = payload["pressure"] if ('pressure' in payload) else None
+    fields["accelerationX"]             = payload["acceleration_x"] if ('acceleration_x' in payload) else None
+    fields["accelerationY"]             = payload["acceleration_y"] if ('acceleration_y' in payload) else None
+    fields["accelerationZ"]             = payload["acceleration_z"] if ('acceleration_z' in payload) else None
+    fields["batteryVoltage"]            = payload["battery"]/1000.0 if hasattr(payload, 'battery') else None
+    fields["txPower"]                   = payload["tx_power"] if ('tx_power' in payload) else None
+    fields["movementCounter"]           = payload["battery"] if ('battery' in payload) else None
+    fields["measurementSequenceNumber"] = payload["measurement_sequence_number"] if ('measurement_sequence_number' in payload) else None
+    fields["tagID"]                     = payload["tagID"] if ('tagID' in payload) else None
+    fields["rssi"]                      = payload["rssi"] if ('rssi' in payload) else None
     return {
-        "measurement": "ruuvitag",
+        "measurement": "ruuvi_measurements",
         "tags": {
-            "mac": mac
+            "mac": mac,
+            "dataFormat": dataFormat
         },
-        "fields": {
-            "temperature": payload["temperature"],
-            "humidity": payload["humidity"],
-            "pressure": payload["pressure"]
-        }
+        "fields": fields
     }
 
 
-client = InfluxDBClient(host="localhost", port=8086, database="tag_data")
+client = InfluxDBClient(host="localhost", port=8086, database="ruuvi")
 
 while True:
     datas = RuuviTagSensor.get_data_for_sensors()