+++ ruuvitag_sensor/decoder.py
@@ -16,10 +16,13 @@ def get_decoder(data_type):
     '''
     if data_type == 2:
         return UrlDecoder()
+    elif data_type == 4:
+        return UrlDecoder()
     elif data_type == 3:
         return Df3Decoder()
     else:
         return Df5Decoder()
+
 
 def twos_complement(value, bits):
     if (value & (1 << (bits - 1))) != 0:
@@ -45,7 +48,7 @@ class UrlDecoder(object):
     Decoder operations are ported from:
     https://github.com/ruuvi/sensor-protocol-for-eddystone-url/blob/master/index.html
 
-    0:   uint8_t     format;          // (0x01 = realtime sensor readings)
+    0:   uint8_t     format;          // (0x02 = realtime sensor readings)
     1:   uint8_t     humidity;        // one lsb is 0.5%
     2-3: uint16_t    temperature;     // Signed 8.8 fixed-point notation.
     4-5: uint16_t    pressure;        // (-50kPa)
@@ -80,11 +83,14 @@ class UrlDecoder(object):
         '''
         try:
             identifier = None
+            data_format = 2
             if len(encoded) > 8:
+                data_format = 4
                 identifier = encoded[8:]
                 encoded = encoded[:8]
             decoded = bytearray(base64.b64decode(encoded, '-_'))
             return {
+                'data_format': data_format,
                 'temperature': self._get_temperature(decoded),
                 'humidity': self._get_humidity(decoded),
                 'pressure': self._get_pressure(decoded),
@@ -141,6 +147,7 @@ class Df3Decoder(object):
             byte_data = data
             acc_x, acc_y, acc_z = self._get_acceleration(byte_data)
             return {
+                'data_format': 3,
                 'humidity': self._get_humidity(byte_data),
                 'temperature': self._get_temperature(byte_data),
                 'pressure': self._get_pressure(byte_data),
@@ -200,7 +207,7 @@ class Df5Decoder(object):
     def _get_powerinfo(self, data):
         '''Return battery voltage and tx power '''
         power_info = (data[13] & 0xFF) << 8 | (data[14] & 0xFF)
-        battery_voltage = rshift(power_info, 5) / 1000 + 1.6
+        battery_voltage = rshift(power_info, 5) + 1600
         tx_power = (power_info & 0b11111) * 2 - 40
 
         if rshift(power_info, 5) == 0b11111111111:
@@ -241,6 +248,7 @@ class Df5Decoder(object):
             byte_data = data
             acc_x, acc_y, acc_z = self._get_acceleration(byte_data)
             return {
+                'data_format': 5,
                 'humidity': self._get_humidity(byte_data),
                 'temperature': self._get_temperature(byte_data),
                 'pressure': self._get_pressure(byte_data),